+++ graphein/ml/datasets/__init__.py
@@ -1,5 +1,10 @@
-from .torch_geometric_dataset import (
-    InMemoryProteinGraphDataset,
-    ProteinGraphDataset,
-    ProteinGraphListDataset,
-)
+from .pdb_data import PDBManager
+
+try:
+    from .torch_geometric_dataset import (
+        InMemoryProteinGraphDataset,
+        ProteinGraphDataset,
+        ProteinGraphListDataset,
+    )
+except (NameError, ImportError):
+    pass