+++ graphein/protein/tensor/testing.py
@@ -4,7 +4,7 @@ from typing import List, Tuple, Union
 
 from loguru import logger as log
 
-from graphein.utils.utils import import_message
+from graphein.utils.dependencies import import_message
 
 from ..resi_atoms import (
     ATOM_NUMBERING_MODIFIED,