+++ tests/ml/test_conversion.py
@@ -12,6 +12,7 @@ from graphein.protein.graphs import construct_graph
 
 try:
     import torch_geometric
+    from torch_geometric.utils import is_undirected
 
     PYG_AVAIL = True
 except ImportError:
@@ -82,3 +83,6 @@ def test_nx_to_pyg(pdb_code):
         data.edge_index.shape[1]
         == data.edge_index_inter.shape[1] + data.edge_index_intra.shape[1]
     )
+
+    # Directed/undirected consistency
+    assert g.is_directed() is not is_undirected(data.edge_index)